+++ src/hooks/useScrollLock.ts
@@ -1,20 +1,18 @@
 "use client";
 
 import React from "react";
-import useId from "hooks/useElementId";
 import { lockScroll, unlockScroll } from "utilities/scroll";
 
 const useScrollLock = () => {
 	const [locked, setLocked] = React.useState(false);
-	const id = useId();
 
 	const handleLockScroll = React.useCallback(() => {
-		lockScroll(id, () => setLocked(true));
-	}, [id]);
+		lockScroll(() => setLocked(true));
+	}, []);
 
 	const handleUnlockScroll = React.useCallback(() => {
-		unlockScroll(id, () => setLocked(false));
-	}, [id]);
+		unlockScroll(() => setLocked(false));
+	}, []);
 
 	return React.useMemo(
 		() => ({